+++ R/RelationalDataConnector_constructors.R
@@ -474,3 +474,126 @@ snowflake_data <- function(..., connection) {
   )
   return(x)
 }
+
+
+
+#' \code{RelationalDataConnector} connector for \code{CDSE}
+#'
+#' @description `r lifecycle::badge("experimental")`
+#' Build data connector for \code{CDSE} datasets
+#'
+#' @export
+#'
+#' @param ... (\code{DatasetConnector} objects)\cr
+#'  dataset connectors created using \code{\link{cdse_dataset_connector}}
+#' @param connection (\code{DataConnection}) object returned from \code{cdse_connection}.
+#'
+#' @return An object of class \code{RelatonalDataConnector}
+#'
+#' @examples
+#'
+#' \dontrun{
+#' x <- cdse_data(
+#'   cdse_cdisc_dataset_connector("ADSL", "cid1234567890"),
+#'   cdse_cdisc_dataset_connector("ADLB", "cid0987654321")
+#' )
+#' app <- init(
+#'   data = cdisc_data(x),
+#'   modules = root_modules(
+#'     module(
+#'       "ADSL AGE histogram",
+#'       server = function(input, output, session, datasets) {
+#'         output$hist <- renderPlot({
+#'           hist(datasets$get_data("ADSL", filtered = TRUE)$AGE)
+#'         })
+#'       },
+#'       ui = function(id, ...) {ns <- NS(id); plotOutput(ns('hist'))},
+#'       filters = "ADSL"
+#'     )
+#'   ),
+#'   header = tags$h1("Sample App")
+#' )
+#' shinyApp(app$ui, app$server)
+#' }
+cdse_data <- function(..., connection = cdse_connection()) {
+  connectors <- list(...)
+  stopifnot(is_class_list("DatasetConnector")(connectors))
+
+  x <- if (any(vapply(connectors, is, logical(1), "CDISCDatasetConnector"))) {
+    CDISCDataConnector$new(connection = connection, connectors = connectors)
+  } else {
+    RelationalDataConnector$new(connection = connection, connectors = connectors)
+  }
+
+  x$set_check(FALSE)
+
+  x$set_ui(
+    function(id, connection, connectors) {
+      ns <- NS(id)
+      div(
+        div(
+          h1("TEAL - Access data on CDSE"),
+          br(),
+          h5("Data access requested for:"),
+          fluidRow(
+            column(
+              11,
+              offset = 1,
+              lapply(seq_along(connectors), function(i) {
+                tags$li(code(connectors[[i]]$get_dataname()),
+                        ": ",
+                        code(connectors[[i]]$get_pull_args()$dataset))
+              })
+            )
+          )
+        ),
+        br(),
+        connection$get_open_ui(ns("open_connection"))
+      )
+    }
+  )
+
+  x$set_server(
+    function(input, output, session, connectors, connection) {
+      # opens connection
+      if (!is.null(connection$get_open_server())) {
+        callModule(connection$get_open_server(),
+                   id = "open_connection",
+                   connection = connection)
+      }
+
+      if (connection$is_opened()) {
+        conn <- connection$get_conn()
+        if (!is.null(conn)) {
+          for (connector in connectors) {
+            connector$get_pull_callable()$assign_to_env("conn", conn)
+          }
+        }
+
+        # call connectors$pull
+        for (connector in connectors) {
+          callModule(connector$get_server(), id = connector$get_dataname())
+          if (connector$is_failed()) {
+            break
+          }
+        }
+
+        if (!is.null(connection$get_close_server())) {
+          callModule(connection$get_close_server(),
+                     id = "close_connection",
+                     connection = connection)
+        }
+      }
+    }
+  )
+
+  x$set_preopen_server(
+    function(input, output, session, connection) {
+      callModule(connection$get_preopen_server(),
+                 id = "open_connection",
+                 connection = connection)
+    }
+  )
+
+  return(x)
+}